# ENH: sparse: efficient conversion from DIA to CSR format