--- a/scipy/sparse/_dia.py
+++ b/scipy/sparse/_dia.py
@@ -428,58 +428,6 @@ class _dia_base(_data_matrix):
 
     tocsr.__doc__ = _spbase.tocsr.__doc__
 
-    def tocsc(self, copy=False):
-        if self.nnz == 0:
-            return self._csc_container(self.shape, dtype=self.dtype)
-
-        num_rows, num_cols = self.shape
-        num_offsets, offset_len = self.data.shape
-        offset_inds = np.arange(offset_len)
-
-        row = offset_inds - self.offsets[:,None]
-        mask = (row >= 0)
-        mask &= (row < num_rows)
-        mask &= (offset_inds < num_cols)
-        mask &= (self.data != 0)
-
-        idx_dtype = self._get_index_dtype(maxval=max(self.shape))
-        indptr = np.zeros(num_cols + 1, dtype=idx_dtype)
-        indptr[1:offset_len+1] = np.cumsum(mask.sum(axis=0)[:num_cols])
-        if offset_len < num_cols:
-            indptr[offset_len+1:] = indptr[offset_len]
-        indices = row.T[mask.T].astype(idx_dtype, copy=False)
-        data = self.data.T[mask.T]
-        return self._csc_container((data, indices, indptr), shape=self.shape,
-                                   dtype=self.dtype)
-
-    tocsc.__doc__ = _spbase.tocsc.__doc__
-
-    def tocoo(self, copy=False):
-        num_rows, num_cols = self.shape
-        num_offsets, offset_len = self.data.shape
-        offset_inds = np.arange(offset_len)
-
-        row = offset_inds - self.offsets[:,None]
-        mask = (row >= 0)
-        mask &= (row < num_rows)
-        mask &= (offset_inds < num_cols)
-        mask &= (self.data != 0)
-        row = row[mask]
-        col = np.tile(offset_inds, num_offsets)[mask.ravel()]
-        idx_dtype = self._get_index_dtype(
-            arrays=(self.offsets,), maxval=max(self.shape)
-        )
-        row = row.astype(idx_dtype, copy=False)
-        col = col.astype(idx_dtype, copy=False)
-        data = self.data[mask]
-        # Note: this cannot set has_canonical_format=True, because despite the
-        # lack of duplicates, we do not generate sorted indices.
-        return self._coo_container(
-            (data, (row, col)), shape=self.shape, dtype=self.dtype, copy=False
-        )
-
-    tocoo.__doc__ = _spbase.tocoo.__doc__
-
     # needed by _data_matrix
     def _with_data(self, data, copy=True):
         """Returns a matrix with the same sparsity structure as self,