--- a/scipy/sparse/_generate_sparsetools.py
+++ b/scipy/sparse/_generate_sparsetools.py
@@ -110,6 +110,7 @@ coo_matvec          v lIITT*T
 coo_matvec_nd       v llIITT*T
 coo_matmat_dense    v llIITT*T
 coo_matmat_dense_nd v lllIIITT*T
+dia_tocsr           v iiiiITI*V*V*W
 dia_matmat          v iiiiITiiiIT*V*W
 dia_matvec          v iiiiITT*T
 dia_matvecs         v iiiiITiT*T