<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
https://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
https://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message.
However, please only include an issue number in the description, not the title,
and please ensure that any code names containing underscores are enclosed in backticks.

Depending on your changes, you can skip CI operations and save time and energy: 
https://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
<!--Example: Closes gh-WXYZ.-->
Addendum to gh-21180, should close gh-21039.

#### What does this implement/fix?
Direct and efficient conversion from DIA format to CSR format, to speed-up all operations not directly implemented for DIA.

#### Additional information
After merging gh-21180, almost all arithmetic operations with DIA arrays became faster than conversion from DIA to other formats, needed for other operations, such as arithmetic with arrays in other formats and passing DIA arrays to methods not implemented for them directly.

For comparison — arithmetic operations:
```
sparse.Arithmetic.time_arithmetic                                                                                                                         ok
======== ==== ============= ============= ============= =============
--                                       op                          
------------- -------------------------------------------------------
 format   XY     __add__       __sub__       multiply      __mul__   
======== ==== ============= ============= ============= =============
  dia     AA     600±20μs      617±20μs    1.34±0.05ms    6.18±0.2ms 
  dia     AB   2.18±0.01ms   2.92±0.02ms   1.33±0.04ms     12.8±2ms  
  dia     BA   2.22±0.04ms   2.21±0.05ms   1.28±0.04ms     12.7±2ms  
  dia     BB   1.26±0.01ms   1.32±0.02ms   3.17±0.04ms    26.5±0.3ms 
======== ==== ============= ============= ============= =============
```
and conversions:
```
sparse.Conversion.time_conversion                                                                                                                         ok
============= ============= ============= ============= ============= ============ ============ =============
--                                                       to_format                                           
------------- -----------------------------------------------------------------------------------------------
 from_format       csr           csc           coo           dia          lil          dok           bsr     
============= ============= ============= ============= ============= ============ ============ =============
     dia       1.74±0.05ms     866±40μs      691±10μs      212±5ns     19.8±0.5ms   22.4±0.3ms   2.28±0.06ms 
============= ============= ============= ============= ============= ============ ============ =============
```
(keep in mind that conversion benchmarks use $100^2 \times 100^2$ arrays, while arithmetic benchmarks are for $250^2 \times 250^2$, and generally with more diagonals, so the relative performance of conversions is even worse than these benchmarks show). This is because currently direct conversion is implemented only to CSC and COO, which is quite inefficient.

The direct and efficient conversion to CSR format implemented in this PR gives it ~5-fold speed-up (which also significantly improves conversion to BSR and slightly to LIL and DOK), making it no longer much slower than arithmetic operations:
```
sparse.Conversion.time_conversion                                                                                                                         ok
============= ============= ============= ============= ============= ============ ============ =============
--                                                       to_format                                           
------------- -----------------------------------------------------------------------------------------------
 from_format       csr           csc           coo           dia          lil          dok           bsr     
============= ============= ============= ============= ============= ============ ============ =============
     dia         273±2μs       814±5μs       597±8μs       217±9ns     17.4±0.1ms   20.5±0.07ms     781±2μs   
============= ============= ============= ============= ============= ============ ============ =============
```

<details>
<summary>P.S.</summary>

* All the benchmarks above were done without the `--quick` option for `asv`, for better reproducibility and more meaningful numbers.
* When I tried to use the `--compare main` approach, it apparently started rebuilding the whole SciPy instead of just recompiling the changed files. Am I missing something?
</details>