@MikhailRyazanov 2025-05-18
The failed tests seem to be failing due to some completely unrelated network problems; there is nothing I can do, but perhaps maintainers can check what is wrong with them and rerun.