@MikhailRyazanov 2025-05-18
By the way, simply removing the current Python + NumPy implementations of `.tocsc()` and `.tocoo()` methods, thus making these conversions to be done indirectly as DIA → CSR → CSC/COO in fact speeds them up by a factor of ~2:
```
============= ============= ============= ============= ============= ============= ============= =============
--                                                        to_format                                            
------------- -------------------------------------------------------------------------------------------------
 from_format       csr           csc           coo           dia           lil           dok           bsr     
============= ============= ============= ============= ============= ============= ============= =============
     dia         270±2μs       484±7μs       386±3μs       207±2ns      17.2±0.2ms    21.2±0.1ms     777±5μs   
============= ============= ============= ============= ============= ============= ============= =============
```

However, [`TestDIA::test_tocoo_gh10050`](https://github.com/scipy/scipy/blob/7ef6aeb632f4d99078d1332108455c401a11da10/scipy/sparse/tests/test_base.py#L5032-L5037) starts failing because the converted COO matrix has `.has_canonical_format = False` even though it is in fact canonical. This seems to be caused by the COO constructor indiscriminately setting
https://github.com/scipy/scipy/blob/7ef6aeb632f4d99078d1332108455c401a11da10/scipy/sparse/_coo.py#L63
even for canonical data. So there are two possible solutions:
1. Relax the test such that it fails only for unsafe results (meaning that `m.has_canonical_format == False` is safe when `inds_are_sorted == True`, like it is implied by the COO constructor, but the opposite is an error).
2. Explicitly add `tocoo()` to `_csr.py` which will transfer `has_canonical_format` to the conversion result (as far as I understand, canonical CSR always produces canonical COO).