@MikhailRyazanov 2025-05-18
While implementing the second “solution” from the previous comment, I've discovered missing bound checks for ill-formed DIA arrays (when diagonal are longer than the array width or lie completely outside), so this is corrected in [677f10b](https://github.com/scipy/scipy/pull/23009/commits/677f10b963052f94aa3ab82af618b3d11a6e189a).

The remaining new commits are related to eliminating the inefficient `tocsc()` and `tocoo()` methods and then dealing with the revealed deficiencies in COO construction. I'm not sure that the last change ([1d03664](https://github.com/scipy/scipy/pull/23009/commits/1d03664c9f24d1da82887e6ddeed32ae46ff3b3b)) is completely correct, but it seems to be more appropriate than setting the default `copy=True` in all `tocoo()` methods (or assuming that internal/intermediate operations will copy the relevant data for methods with default `copy=False`).