@MikhailRyazanov 2025-05-19
Here are more extensive benchmarks for arrays of various sizes and comparing conversions with arithmetic operations on equal footing. `main:` means how it is now, `this:` means this PR; arithmetic benchmarks are not affected.

Using `poisson2d(20, ...)` = 400 × 400 array:
```
      sparse.Conversion.time_conversion
      --                                      to_format
      ------------- ---------------------------------------------------------
       from_format      csr        csc       coo      lil     dok      bsr   
      ============= =========== ========= ========= ======= ======= =========
main:      dia      225±2μs     122±10μs  117±0.3μs 882±3μs 721±5μs 296±0.3μs
this:      dia       57.0±0.3μs 104±0.7μs 111±0.3μs 648±5μs 522±2μs 114±0.3μs

sparse.Arithmetic.time_arithmetic
--                              op
------------- ---------------------------------------
 format   XY     __add__      __sub__      multiply  
======== ==== ============= ============ ============
  dia     AA   43.0±0.5μs    43.3±0.4μs   64.9±0.5μs  
```

Using `poisson2d(100, ...)` = 10000 × 10000 array:
```
      sparse.Conversion.time_conversion
      --                                    to_format
      ------------- ---------------------------------------------------------------
       from_format      csr       csc      coo       lil        dok         bsr    
      ============= =========== ======== ======== ========== ========== ===========
main:     dia       1.74±0.05ms 866±40μs 691±10μs 19.8±0.5ms 22.4±0.3ms 2.28±0.06ms
this:     dia         270±2μs   484±7μs  386±3μs  17.2±0.2ms 21.2±0.1ms   777±5μs  

sparse.Arithmetic.time_arithmetic
--                               op
------------- -----------------------------------------
 format   XY     __add__       __sub__       multiply  
======== ==== ============= ============= =============
  dia     AA    99.8±0.2μs    101±0.3μs      168±1μs   
```

Using `poisson2d(250, ...)` = 62500 × 62500 array:
```
      sparse.Conversion.time_conversion
      --                                        to_format
      ------------- ----------------------------------------------------------------
       from_format     csr         csc        coo        lil       dok        bsr   
      ============= =========== ========== ========== ========= ========= ==========
main:      dia      15.5±0.6ms  7.31±0.6ms 8.89±1ms   126±1ms   142±1ms   19.3±0.6ms
this:      dia       4.55±0.3ms 6.65±0.1ms 5.24±0.3ms 114±0.7ms 130±0.4ms 7.99±0.2ms

sparse.Arithmetic.time_arithmetic
--                               op
------------- -----------------------------------------
 format   XY     __add__       __sub__       multiply  
======== ==== ============= ============= =============
  dia     AA     600±20μs      617±20μs    1.34±0.05ms 
```

So conversion from DIA becomes 3–6 times faster to CSR, >2 times faster to BSR, and somewhat faster to all other formats. Still slower than simple arithmetic operations, but not as much as it currently is (like an order of magnitude for the 10000 × 10000 size).