@MikhailRyazanov 2025-06-21
@dschult, @perimosocordiae:
OK, here is the improved variant, with the output arrays allocated by NumPy, based on `nnz`, and then pruned. And hopefully also addressing all the review comments above.

Benchmark comparisons for various sizes (see [above](https://github.com/scipy/scipy/pull/23009#issuecomment-2889491978)); “`main`” is the current implementation from the main branch, “`this`” is this PR (other lines are explained below):
```
          sparse.Conversion.time_conversion
          --                                                 to_format
          ------------- ------------------------------------------------------------------------------------
           from_format        csr           csc          coo          lil           dok            bsr
          ============= ============== ============= ============ ============= ============= ==============
400 × 400:
  main:        dia       225±2μs        122±10μs      117±0.3μs    882±3μs       721±5μs       296±0.3μs
  this:        dia        71.5±0.09μs   122±0.5μs     123±0.3μs    661±1μs       544±6μs       131±0.1μs

  no nnz:      dia        51.9±2μs       97.5±0.3μs   106±0.8μs    635±3μs       511±1μs       109±0.5μs
  +copy:       dia        54.1±0.2μs    101±0.7μs     107±0.6μs    644±2μs       518±2μs       111±0.3μs

  cntnz:       dia        50.4±0.2μs     99.5±0.3μs   108±1μs      644±5μs       565±60μs      110±1μs

10000 × 10000:
  main:        dia       1.74±0.05ms    866±40μs      691±10μs     19.8±0.5ms    22.4±0.3ms    2.28±0.06ms
  this:        dia         175±1μs      392±4μs       290±7μs      17.2±0.08ms   18.8±0.1ms      692±8μs

  no nnz:      dia         157±0.5μs    366±3μs       266±2μs      16.9±0.08ms   18.6±0.06ms     662±3μs
  +copy:       dia         198±1μs      417±3μs       318±3μs      17.3±0.07ms   18.8±0.09ms     714±7μs

  cntnz:       dia         195±1μs      421±10μs      305±0.7μs    17.1±0.1ms    18.6±0.2ms      695±7μs

62500 × 62500:
  main:        dia       15.5±0.6ms     7.31±0.6ms    8.89±1ms     126±1ms       142±1ms       19.3±0.6ms
  this:        dia        1.06±0.02ms   5.94±0.07ms   3.05±0.6ms   113±0.5ms     126±0.3ms      6.80±0.06ms

  no nnz:      dia        1.07±0.04ms   5.76±0.08ms   3.66±0.9ms   113±1ms       127±0.5ms      6.81±0.2ms
  +copy:       dia        4.06±0.2ms    6.93±0.5ms    4.66±0.2ms   115±0.9ms     130±0.6ms      7.76±0.3ms

  cntnz:       dia        1.38±0.04ms   6.24±0.2ms    3.50±0.9ms   113±0.7ms     131±0.8ms      7.13±0.2ms
```

For small sizes, the performance is still better than the original but slightly worse than from the previous attempt with `nnz` calculated during the conversion (likely because `._getnnz()` adds a constant, size-independent, overhead), however, for larger sizes, the performance gain is now even better than before.

I've also tried two alternative approaches:
1. Using `max_nnz = self.data.size` instead of `max_nnz = self.nnz`, labeled “`no nnz`” in the benchmarks above. This improves the timings by ~20 μs, which is noticeable for small sizes, but at the same time is more likely to trigger copying when pruning the output, with a huge impact for large sizes (see the “`+copy`” lines above, which include artificially triggering it), plus higher memory usage (might be very undesirable when sizes are large).
2. Using `self.count_nonzero()` to get the exact output size before the conversion and thus avoid any memory overhead and reallocations. This would be useless with the current `count_nonzero()` implementation, which is slower than the conversion itself, but reimplementing it in C++ (see the [dia-cntnz](https://github.com/MikhailRyazanov/scipy/tree/dia-cntnz) branch in my fork) gives ~7-fold speed-up, and the overall results are shown as “`cntnz`” above. They are quite good for small sizes, but for large sizes are ~25% slower than “`this`”, although still ~10 faster than “`main`” and cannot degrade to the worst-case “`+copy`”. However, another C++ function (`dia_count_nonzero()`) additionally increases the library size by about as much as this PR does (with only added `dia_tocsr()`), so using this approach would be reasonable only if fast `count_nonzero()` for DIA arrays is useful on its own.

I don't think that the first alternative is really better, despite its slightly better best-case performance. The second alternative has some advantages and some disadvantages, so I'm not sure. What do you think?